refactor: split getElectronVersion as a module
