+++ packages/@wdio_electron-utils/test/index.spec.ts
@@ -2,8 +2,8 @@ import path from 'node:path';
 import { expect, it, vi, describe, beforeEach } from 'vitest';
 import fs from 'node:fs/promises';
 
-import { getBinaryPath, getAppBuildInfo, getElectronVersion } from '../src/index.js';
-import { NormalizedPackageJson, NormalizedReadResult } from 'read-package-up';
+import { getBinaryPath, getAppBuildInfo } from '../src/index.js';
+import { NormalizedPackageJson } from 'read-package-up';
 import { AppBuildInfo } from '@wdio/electron-types';
 
 // Mock our own readConfig implementation for testing
@@ -83,11 +83,6 @@ function getBuilderConfigCandidates(configFileName = 'electron-builder'): string
 
 function getFixturePackagePath(moduleType: string, fixtureName: string) {
   return path.resolve(process.cwd(), '..', '..', 'fixtures', moduleType, fixtureName, 'package.json');
-}
-
-// Utility function to get the directory from a package.json path
-function getProjectDirFromPackagePath(packagePath: string): string {
-  return path.dirname(packagePath);
 }
 
 // Utility function to check if a path includes the fixture type
@@ -544,329 +539,6 @@ beforeEach(() => {
 });
 
 describe('Electron Utilities', () => {
-  describe('getElectronVersion()', () => {
-    it('should return the electron version from package.json dependencies', async () => {
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          dependencies: {
-            electron: '25.0.1',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: '/path/to/package.json',
-      } as NormalizedReadResult;
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('25.0.1');
-    });
-
-    it('should return the electron version from package.json devDependencies', async () => {
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            electron: '25.0.1',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: '/path/to/package.json',
-      } as NormalizedReadResult;
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('25.0.1');
-    });
-
-    it('should return the nightly electron version from package.json dependencies', async () => {
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          dependencies: {
-            'electron-nightly': '33.0.0-nightly.20240621',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: '/path/to/package.json',
-      } as NormalizedReadResult;
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('33.0.0-nightly.20240621');
-    });
-
-    it('should return the nightly electron version from package.json devDependencies', async () => {
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron-nightly': '33.0.0-nightly.20240621',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: '/path/to/package.json',
-      } as NormalizedReadResult;
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('33.0.0-nightly.20240621');
-    });
-
-    it('should return undefined when there is no electron dependency', async () => {
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          dependencies: {},
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: '/path/to/package.json',
-      } as NormalizedReadResult;
-      const version = await getElectronVersion(pkg);
-      expect(version).toBeUndefined();
-    });
-
-    it('should fetch the electron version from pnpm workspace default catalog', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalog: {
-          electron: '29.4.1',
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalog:\n  electron: "29.4.1"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            electron: 'catalog:',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('29.4.1');
-      // Since we're traversing directories, we don't need to check the exact path anymore
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should fetch the electron-nightly version from pnpm workspace default catalog', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalog: {
-          'electron-nightly': '33.0.0-nightly.20240621',
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalog:\n  electron-nightly: "33.0.0-nightly.20240621"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron-nightly': 'catalog:',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('33.0.0-nightly.20240621');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should fetch the electron version from pnpm workspace named catalog', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalogs: {
-          stable: {
-            electron: '29.4.1',
-          },
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalogs:\n  stable:\n    electron: "29.4.1"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            electron: 'catalog:stable',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('29.4.1');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should fetch the electron-nightly version from pnpm workspace named catalog', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalogs: {
-          nightly: {
-            'electron-nightly': '33.0.0-nightly.20240621',
-          },
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalogs:\n  nightly:\n    electron-nightly: "33.0.0-nightly.20240621"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron-nightly': 'catalog:nightly',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('33.0.0-nightly.20240621');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should prioritize electron over electron-nightly when both are using catalogs', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalog: {
-          'electron': '29.4.1',
-          'electron-nightly': '33.0.0-nightly.20240621',
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalog:\n  electron: "29.4.1"\n  electron-nightly: "33.0.0-nightly.20240621"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron': 'catalog:',
-            'electron-nightly': 'catalog:',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('29.4.1');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it("should fallback to electron-nightly when electron catalog doesn't exist", async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalog: {
-          'electron-nightly': '33.0.0-nightly.20240621',
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve('catalog:\n  electron-nightly: "33.0.0-nightly.20240621"');
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron': 'catalog:',
-            'electron-nightly': 'catalog:',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('33.0.0-nightly.20240621');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should handle a mix of named and default catalogs', async () => {
-      vi.mocked(await import('yaml')).parse.mockReturnValueOnce({
-        catalog: {
-          'electron-nightly': '33.0.0-nightly.20240621',
-        },
-        catalogs: {
-          stable: {
-            electron: '29.4.1',
-          },
-        },
-      });
-
-      const pkgPath = '/path/to/project/package.json';
-
-      // Set up mock to succeed with the workspace file
-      vi.mocked(fs.readFile).mockImplementationOnce((_path) => {
-        return Promise.resolve(
-          'catalog:\n  electron-nightly: "33.0.0-nightly.20240621"\ncatalogs:\n  stable:\n    electron: "29.4.1"',
-        );
-      });
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            'electron': 'catalog:stable',
-            'electron-nightly': 'catalog:',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBe('29.4.1');
-      expect(fs.readFile).toHaveBeenCalled();
-    });
-
-    it('should gracefully handle missing workspace file', async () => {
-      vi.mocked(fs.readFile).mockRejectedValueOnce(new Error('File not found'));
-
-      const pkgPath = '/path/to/project/package.json';
-      const projectDir = getProjectDirFromPackagePath(pkgPath);
-
-      const pkg = {
-        packageJson: {
-          name: 'my-app',
-          version: '1.0.0',
-          devDependencies: {
-            electron: 'catalog:stable',
-          },
-        } as Omit<NormalizedPackageJson, 'readme' | '_id'>,
-        path: pkgPath,
-      } as NormalizedReadResult;
-
-      const version = await getElectronVersion(pkg);
-      expect(version).toBeUndefined();
-      expect(fs.readFile).toHaveBeenCalledWith(path.join(projectDir, 'pnpm-workspace.yaml'), 'utf8');
-    });
-  });
-
   describe('getAppBuildInfo()', () => {
     describe('ESM', () => {
       it('should throw an error when no build tools are found', async () => {
